requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
fastcore==1.6.9
fastlite==0.0.7
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
msal==1.30.0
msal-extensions==1.2.0
multidict==6.1.0
numpy==2.0.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
playwright==1.45.1
pluggy==1.5.0
portalocker==2.10.1
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.3.6
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlite_minutils==3.36.0.post4
starlette==0.38.2
text-unidecode==1.3
trio==0.26.0
trio-websocket==0.11.1
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.23.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
yarl==1.11.1
